Inject scrape-run timeout from composition root; move TimeoutError to shared/errors
The application-layer use case read process.env directly — the only such read in the layer. Read env in the composition root and inject via the existing runTimeoutMs dep, mirroring pollIntervalMs. TimeoutError now lives with the other shared errors (kept as a plain Error, not a DomainError — it's control-flow, not an HTTP fault).
