+++ src/shared/errors/index.ts
@@ -4,3 +4,4 @@ export { ValidationError } from './ValidationError.js'
 export { ConflictError } from './ConflictError.js'
 export { UnauthorizedError } from './UnauthorizedError.js'
 export { ForbiddenError } from './ForbiddenError.js'
+export { TimeoutError } from './TimeoutError.js'
